requirements.txt
requests
django
pytest
pytest-django
djangorestframework
djangorestframework-simplejwt
django-filter
django-cors-headers